requirements.txt
cachetools==5.3.3
discord.py==2.3.2
Flask==3.0.3
Flask-Discord==0.1.69
matplotlib==3.8.3
pandas==2.2.1
requests-oauthlib==2.0.0
python-dotenv==1.0.0